feat: move Caddy Reverse Proxy to main navigation as /proxy page
- Create new top-level /proxy page with full proxy host management UI
  (list, add, edit, delete, toggle, search, sync, test connection)
- Add "Proxy" entry to sidebar navigation with Shield icon
- Simplify Settings > Caddy page to only show Admin API URL config
- Update settings nav description to reflect reduced scope

Closes #397
